requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Django==4.0
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
httpie==2.6.0
idna==3.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
psycopg2==2.9.2
Pygments==2.10.0
PyMySQL==1.0.2
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
torch==1.10.0
torchvision==0.11.1
typing_extensions==4.0.0
urllib3==1.26.7
websocket==0.2.1
websockets==10.1
zope.event==4.5.0
zope.interface==5.4.0
